improve error message and allow non-existing files for --rex-authpath
Users might pass a folder, but we really expect a file here, so let the
error message say this.

Also, the code can actually handle a non-existing file just fine, as it
will create it if it does not exist, so let's just error if the path
given exists but is not a file.

Fixes: #296