BZ #1659414 - don't fail on unknown smart proxies
In some scenarios (load balanced or multi homed smart proxy), the name
of the proxy as seen by the client is not known to Foreman/Katello.

Instead of failing, inform the user that the name could not be used for
enforcing the content source/Puppet/OpenSCAP proxy and carry on.