don't fail when DNF can't reach repos during sub-man prereqs
When a host has repositories configured it can't reach, DNF will error
out when we call `dnf_base.fill_sack()` and abort the execution of the
script.
However, we only need repo data to learn whether sub-man is available,
but given in most cases it's installed already, we don't need this
information and can live with it not being upgradeable (the code later
on handles that case gracefully).